Drawer navigation
Drawer navigation

All buttons point to something, if the page is not built up it has sent to the dashboard,

As of right now there is not a way to go back from a button press, as in each is a one way trip.

There is a fix but it throws an error, I will update the trello.